legal-document-generator/requirements.txt
boto3==1.28.79
botocore==1.31.79
fpdf
streamlit==1.30.0
python-dotenv==1.0.0
python-docx